It was developed from the crud of the doctors

+++ src/app/pages/maintenance/hospitals/hospitals.component.ts
@@ -76,9 +76,21 @@ export class HospitalsComponent implements OnInit, OnDestroy {
   }
 
   public deleteHospital(hospital: Hospital) {
-    this._hospitalService.deleteHospital(hospital._id).subscribe(() => {
-      this.loadHospitals();
-      Swal.fire('Removed', hospital.name, 'success');
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._hospitalService.deleteHospital(hospital._id).subscribe(() => {
+          this.loadHospitals();
+          Swal.fire('Removed', hospital.name, 'success');
+        });
+      }
     });
   }
